% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game}
\alias{game}
\title{Game}
\format{A data frame with 20 rows and 4 variables:
\describe{
  \item{Round}{Round number}
  \item{Player}{Player's Throw}
  \item{R}{R's Throw}
  \item{Result}{Result for Player either "Win", "Tie" or "Lose"}
}}
\usage{
win_stats
}
\description{
This data contains sample output from a call of paperstone("paper", 20)
}
\keyword{datasets}
