% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_pair.R
\name{sps_pair}
\alias{sps_pair}
\title{Returns the paired value/number for Scissors-Paper-Stone.}
\usage{
sps_pair(x)
}
\arguments{
\item{x}{A throw}
}
\value{
The corresponding throw name/integer code
}
\description{
When the input is the integer code, the output is the corresponding throw in words.
When the input is the throw name, the output is the corresponding integer code.
The pairs are "Scissors = 1", "Stone = 2", "Paper = 3"
}
\examples{
sps_pair(1)
sps_pair("stone")
}
