% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_stats.R
\name{win_stats}
\alias{win_stats}
\title{Calculate your results from your Scissors-Paper-Stone game!}
\usage{
win_stats(game, outcome = "Win")
}
\arguments{
\item{game}{Dataframe containing output of paperstone()}

\item{outcome}{Outcome type to be counted, default = "Win"}
}
\value{
The proportion of that outcome in the game.
}
\description{
Calculates the proportion of a certain outcome for a game of Scissors-Paper-Stone!
}
\examples{
win_stats(game)
win_stats(game, "Tie")
}
